tr2/savegame: isolate legacy savegame module
This isolates TR2's legacy savegame logic into its own module, similar
to TR1, and updates related usage throughout.
