+++ src/tr2/game/game_flow/sequencer_misc.c
@@ -1,9 +1,9 @@
-#include "decomp/savegame.h"
 #include "game/game.h"
 #include "game/game_flow/common.h"
 #include "game/game_flow/sequencer.h"
 #include "game/game_flow/vars.h"
 #include "game/level.h"
+#include "game/savegame.h"
 
 #include <libtrx/game/game_string_table.h>
 #include <libtrx/log.h>
