+++ src/tr2/game/level.c
@@ -1,7 +1,6 @@
 #include "game/level.h"
 
 #include "decomp/decomp.h"
-#include "decomp/savegame.h"
 #include "game/camera.h"
 #include "game/effects.h"
 #include "game/game.h"
@@ -16,6 +15,7 @@
 #include "game/random.h"
 #include "game/render/common.h"
 #include "game/room.h"
+#include "game/savegame.h"
 #include "game/shell.h"
 #include "game/sound.h"
 #include "game/stats.h"
