+++ src/tr2/game/option/option_compass.c
@@ -1,8 +1,8 @@
-#include "decomp/savegame.h"
 #include "game/game.h"
 #include "game/input.h"
 #include "game/option/option.h"
 #include "game/requester.h"
+#include "game/savegame.h"
 #include "game/sound.h"
 #include "game/ui/widgets/stats_dialog.h"
 #include "global/vars.h"
