+++ src/tr2/game/option/option_passport.c
@@ -1,5 +1,4 @@
 #include "decomp/decomp.h"
-#include "decomp/savegame.h"
 #include "game/game.h"
 #include "game/game_flow.h"
 #include "game/game_string.h"
@@ -7,6 +6,7 @@
 #include "game/inventory_ring.h"
 #include "game/option/option.h"
 #include "game/requester.h"
+#include "game/savegame.h"
 #include "game/sound.h"
 #include "game/text.h"
 #include "global/vars.h"
