+++ src/tr2/game/savegame/savegame_legacy.h
@@ -1,0 +1,6 @@
+#pragma once
+
+#include <libtrx/filesystem.h>
+
+void Savegame_Legacy_SaveToFile(MYFILE *fp);
+void Savegame_Legacy_LoadFromFile(MYFILE *fp);
