+++ src/tr2/meson.build
@@ -90,7 +90,6 @@ sources = [
   init,
   'decomp/decomp.c',
   'decomp/flares.c',
-  'decomp/savegame.c',
   'decomp/skidoo.c',
   'game/camera.c',
   'game/clock.c',
@@ -267,6 +266,7 @@ sources = [
   'game/room.c',
   'game/room_draw.c',
   'game/savegame/common.c',
+  'game/savegame/savegame_legacy.c',
   'game/scaler.c',
   'game/shell/common.c',
   'game/shell/input.c',
